+++ ev.c
@@ -1433,8 +1433,8 @@ const char* ev_strerror(int err)
 // #line 23 "ev.c"
 ////////////////////////////////////////////////////////////////////////////////
 // FILE:    ev/fs.c
-// SIZE:    24526
-// SHA-256: 8e7f368872ebf219b284dc8d05b017eb3cb1151badc3374e392e694cfe47d49d
+// SIZE:    24939
+// SHA-256: cff6e6766192e053eee20e1293181dfe8b14ab6cc0f9ff69681ec249743f1e1a
 ////////////////////////////////////////////////////////////////////////////////
 // #line 1 "ev/fs.c"
 #include <sys/stat.h>
@@ -2150,6 +2150,13 @@ int64_t ev_file_seek(ev_file_t* file, ev_fs_req_t* req, int whence, int64_t offs
     return 0;
 }
 
+ssize_t ev_file_read(ev_file_t* file, ev_fs_req_t* req, void* buff,
+    size_t size, ev_file_cb cb)
+{
+    ev_buf_t buf = ev_buf_make(buff, size);
+    return ev_file_readv(file, req, &buf, 1, cb);
+}
+
 ssize_t ev_file_readv(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
     size_t nbuf, ev_file_cb cb)
 {
@@ -2172,6 +2179,13 @@ ssize_t ev_file_preadv(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],
     }
 
     return _ev_file_read_template(file, req, bufs, nbuf, offset, cb, _ev_file_on_pread);
+}
+
+ssize_t ev_file_write(ev_file_t* file, ev_fs_req_t* req, const void* data,
+    size_t size, ev_file_cb cb)
+{
+    ev_buf_t buf = ev_buf_make((void*)data, size);
+    return ev_file_writev(file, req, &buf, 1, cb);
 }
 
 ssize_t ev_file_writev(ev_file_t* file, ev_fs_req_t* req, ev_buf_t bufs[],