+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 1. `ev_fs_readdir()` is able to operator in synchronous mode.
 2. `ev_fs_readfile()` is able to operator in synchronous mode.
 3. `ev_file_seek()` is able to operator in synchronous mode.
+4. support normal `ev_file_read()` and `ev_file_write()`.
 
 ### Bug Fixes
 1. `ev_hrtime()` no longer require initialize event loop first.